big org change, minor touchup
