+++ src-player_client/playerclient/playerclient.c
@@ -1,9 +1,4 @@
-#include <unistd.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include "wireguard.h"
-
+#include "playerclient.h"
 #define BUF_SIZE 256
 
 //init the connection and allow user to mess around in shell
@@ -140,50 +135,3 @@ void process_config(const char* file_name, const char* search_str, const char* r
     // attempt to bring up wiregaurd daemon 
 }
 
-
-
-// Casey I know this is NOT secure just getting the logic ready :)
-int main(int argc, char* argv[])
- {
-    char user_input[4096];
-    char pubkey[32], privkey[32];
-    
-
-
-    if (argc < 4) 
-    {
-        printf("Usage: %s [user] [hostname] [shell]\n", argv[0]);
-        return 1;
-    }
-
-    // check for existing config, if not regen
-    if(check_key(pubkey,privkey) == 0)
-        wgkeygen(pubkey,privkey);
-
-
-    
-    // very simple menuing 
-    while(1)
-    {
-        printf("\nPlease enter some text: ");
-        fgets(user_input, sizeof(user_input), stdin);
-
-        if(strcmp(user_input, "connect\n") == 0 )
-            connectclient(argc,argv);
-        
-        else if(strcmp(user_input, "wgkeygen\n")  == 0 )
-            wgkeygen(pubkey, privkey);
-
-        else if(strcmp(user_input, "proccfg\n") == 0 )
-            process_config("file.txt", "<your private key here>", privkey);
-    
-        else if(strcmp(user_input, "stop\n") == 0 )
-            break;
-        
-        else
-            printf("UNKNOWN\n");
-    }
- 
-
-    return 0;
-}
