+++ src-player_client/playershell/shellserver.c
@@ -1,26 +1,10 @@
-#include <stdio.h>
-#include <stdlib.h>
-#include <unistd.h>
-#include <string.h>
-#include <sys/wait.h>
+#include "shellserver.h"
 
 
 #define BUFSIZE 256
 #define INPUTLEN 32
 
-/* 
- * Usernames can only be composed of
- * - lowercase alphabet
- * - numerics
- * - underscore
- */
-const char * const allowed_chars = "abcdefghijklmnopqrstuvwxyz0123456789_";
 
-/* 
- * Records current user that we are running commands for
- * Authenticated by SSH server
- */
-char* session_user = NULL;
 
 void read_credentials(char *username, char *password)
 {
@@ -109,6 +93,16 @@ void retrieve_conf(char* key, char* username, char* password, char* gameid)
     return;
 }
 
+char *append_p(const char *str) {
+    size_t len = strlen(str);
+    char *result = malloc(len + 3);
+    
+    strcpy(result, "-p");
+    strcat(result, str);
+    return result;
+}
+
+
 
 void heartbeat (char* key, char* username, char* password)
 {
@@ -116,10 +110,8 @@ void heartbeat (char* key, char* username, char* password)
     char buf[BUFSIZE];
     char userid[BUFSIZE];
     char gameid[BUFSIZE];
-    char passarg[(strlen(password) + 2)];
+    char* passarg = append_p(password);
 
-    strcat(passarg,"-p");
-    strcat(passarg,password);
 
     //grab user id to search for games
     query = "SELECT, user_id FROM devices, where pubkey = '";
@@ -188,6 +180,8 @@ void heartbeat (char* key, char* username, char* password)
     }
 
     return;
+
+    free(passarg);
 }
 
 
@@ -284,93 +278,6 @@ void run_command (char* c)
     return;
 }
 
-int login (char* c)
-{
-    // copy the argument for tok'ing
-    char* command = (char*) calloc (4096, sizeof (char));
-    memcpy (command, c, strlen (c));
-    // tokenize for (command, arg)
-    char* arg = strtok (command, " ");
-
-    // check that the first command is indeed 'login'
-    // as it should be
-    if (strcmp (arg, "login"))
-    {
-        free (command);
-        return 1;
-    }
-
-    // grab the argument
-    arg = strtok (NULL, " ");
-
-    // if there is no arg (no username to set), then fail
-    if (! arg)
-    {
-        free (command);
-        return 1;
-    }
-
-    // check that the requested username is made up of only allowed characters
-    if (strspn (arg, allowed_chars) < strlen (arg))
-    {
-        free (command);
-        return 1;
-    }
-
-    // copy username into session_user
-    session_user = (char*) calloc (strlen (arg) + 1, sizeof (char));
-    memcpy (session_user, arg, strlen (arg));
-
-    // cleanup
-    free (command);
-    return 0;
-}
-
-void cleanup ()
-{
-    if (session_user)
-        free (session_user);
-}
 
 
-int main (int argc, char* argv[]) 
-{
-    /*
-     * If the ssh server and authentication are
-     * set up correctly, the shell should always
-     * be called non-interactively, and should
-     * run the login command before anything else
-     */
 
-    // non-interactive commands
-    if (argc > 2 && ! strcmp (argv[1], "-c"))
-    {
-        /* 
-         * login() will check that our command is indeed 'login'
-         * returns 0 on success
-         */
-        int r = login (argv[2]);
-        // if we failed to login, cleanup and return
-        if (r)
-        {
-            printf ("Authentication failed.");
-            cleanup ();
-            return 1;
-        }
-        printf ("Hello, %s.\n", session_user);
-
-        /* 
-         * Client would run: ssh host [command] 
-         * This will grab and run the original command sent
-         */
-        run_command (getenv ("SSH_ORIGINAL_COMMAND"));
-    }
-    // interactive
-    else
-    {
-        printf ("Interactive sessions are not allowed. Goodbye!\n");
-    }
-
-    cleanup ();
-    return 0;
-}
